# docs/swagger.yaml
definitions:
...
  database.Namespace:
    properties:
      created_at:
        type: string
      id:
        type: integer
      mirrored:
        type: boolean
      namespace_type:
        $ref: '#/definitions/database.NamespaceType'
      path:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
    type: object
...
  database.Organization:
    properties:
      created_at:
        type: string
      description:
        type: string
      git_path:
        type: string
      homepage:
        type: string
      id:
        type: integer
      industry:
        type: string
      logo:
        type: string
      name:
        type: string
      namespace:
        $ref: '#/definitions/database.Namespace'
      namespace_id:
        type: integer
      org_type:
        type: string
      path:
        description: unique name of the organization
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
...
  types.APIBadRequest:
    type: object
  types.APIInternalServerError:
    type: object
...
  types.Code:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
    type: object
  types.Collection:
    properties:
      avatar:
        type: string
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      likes:
        type: integer
      name:
        type: string
      namespace:
        type: string
      nickname:
        type: string
      private:
        type: boolean
      repositories:
        items:
          $ref: '#/definitions/types.CollectionRepository'
        type: array
      theme:
        type: string
      updated_at:
        type: string
      user_likes:
        type: boolean
      username:
        type: string
    type: object
...
  types.Dataset:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
    type: object
...
  types.DeleteOrgReq:
    type: object
...
  types.EditOrgReq:
    properties:
      description:
        example: org description
        type: string
      homepage:
        example: https://www.example.com
        type: string
      industry:
        type: string
      logo:
        example: https://www.example.com/logo.png
        type: string
      nickname:
        description: Display name
        example: org display name
        type: string
      org_type:
        example: company or school etc
        type: string
      verified:
        example: false
        type: boolean
    type: object
...
  types.Model:
    properties:
      base_model:
        type: string
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      enable_finetune:
        type: boolean
      enable_inference:
        type: boolean
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      status:
        description: url to interact with the model
        example: RUNNING
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
      widget_type:
        allOf:
        - $ref: '#/definitions/types.ModelWidgetType'
        description: 'widget UI style: generation,chat'
        example: generation
    type: object
...
  types.Organization:
    properties:
      description:
        type: string
      homepage:
        type: string
      industry:
        type: string
      logo:
        type: string
      name:
        type: string
      org_type:
        type: string
      path:
        description: unique name of the organization
        type: string
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
...
  types.Response:
    properties:
      data: {}
      msg:
        type: string
    type: object
  types.ResponseWithTotal:
    properties:
      data: {}
      msg:
        type: string
      total:
        type: integer
    type: object
...
  types.Space:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      cover_image_url:
        type: string
      created_at:
        type: string
      default_branch:
        type: string
      description:
        example: ""
        type: string
      endpoint:
        description: the serving endpoint url
        example: https://localhost/spaces/myname/myspace
        type: string
      env:
        type: string
      hardware:
        type: string
      id:
        type: integer
      license:
        example: MIT
        type: string
      like_count:
        type: integer
      name:
        example: space_name_1
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        example: ""
        type: string
      path:
        example: user_or_org_name/space_name_1
        type: string
      private:
        type: boolean
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      sdk:
        description: like gradio,steamlit etc
        example: "1"
        type: string
      sdk_version:
        example: v0.1
        type: string
      secrets:
        type: string
      sku:
        type: string
      source:
        $ref: '#/definitions/types.RepositorySource'
      status:
        description: deploying, running, failed
        type: string
      svc_name:
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      template:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
      username:
        example: creator_user_name
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /organization/{namespace}:
    delete:
      consumes:
      - application/json
      description: delete organization
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the op user
        in: query
        name: current_user
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.DeleteOrgReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Delete organization
      tags:
      - Organization
    get:
      consumes:
      - application/json
      parameters:
      - description: the op user
        in: query
        name: current_user
        type: string
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Organization'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization info
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: update organization
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the op user
        in: query
        name: current_user
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.EditOrgReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/database.Organization'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Update organization
      tags:
      - Organization
  /organization/{namespace}/codes:
    get:
      consumes:
      - application/json
      parameters:
      - description: org name
        in: path
        name: namespace
        required: true
        type: string
      - description: current user name
        in: query
        name: current_user
        required: true
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Code'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization codes
      tags:
      - Organization
  /organization/{namespace}/collections:
    get:
      consumes:
      - application/json
      parameters:
      - description: org name
        in: path
        name: namespace
        required: true
        type: string
      - description: current user name
        in: query
        name: current_user
        required: true
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Collection'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization Collections
      tags:
      - Organization
  /organization/{namespace}/datasets:
    get:
      consumes:
      - application/json
      description: get organization datasets
      parameters:
      - description: org name
        in: path
        name: namespace
        required: true
        type: string
      - description: current user name
        in: query
        name: current_user
        required: true
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Dataset'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization datasets
      tags:
      - Organization
...
  /organization/{namespace}/models:
    get:
      consumes:
      - application/json
      description: get organization models
      parameters:
      - description: org name
        in: path
        name: namespace
        required: true
        type: string
      - description: current user name
        in: query
        name: current_user
        required: true
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Model'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization models
      tags:
      - Organization
  /organization/{namespace}/spaces:
    get:
      consumes:
      - application/json
      parameters:
      - description: org name
        in: path
        name: namespace
        required: true
        type: string
      - description: current user name
        in: query
        name: current_user
        required: true
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Space'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organization Spaces
      tags:
      - Organization
  /organizations:
    get:
      consumes:
      - application/json
      description: get organizations
      parameters:
      - description: org type
        in: query
        name: org_type
        type: string
      - description: industry
        in: query
        name: industry
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: page size
        in: query
        name: per
        type: integer
      - description: current page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Organization'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get organizations
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: create a new organization
      parameters:
      - description: the op user
        in: query
        name: current_user
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateOrgReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Organization'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Create a new organization
      tags:
      - Organization
...
securityDefinitions:
  ApiKey:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"